server was able to be activated in the developement server
