+++ server/app/__init__.py
@@ -4,84 +4,22 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_marshmallow import Marshmallow
 import datetime
 
-app = Flask(__name__)
-CORS(app)
+db = SQLAlchemy()
+ma = Marshmallow()
 
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+def create_app():
+    app = Flask(__name__)
+    CORS(app)
 
-db = SQLAlchemy(app)
-ma = Marshmallow(app)
+    app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite'
+    app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(100))
-    email = db.Column(db.String(100))
-    password = db.Column(db.String(100))
-    created_at = db.Column(db.DateTime, default=datetime.datetime.now)
+    db.init_app(app)
+    ma.init_app(app)
 
-    def __init__(self, name, email, password):
-        self.name = name
-        self.email = email
-        self.password = password
+    with app.app_context():
+        from .routes import user_bp
+        app.register_blueprint(user_bp)
 
-class UserSchema(ma.Schema):
-    class Meta:
-        fields = ('id', 'name', 'email', 'password', 'created_at')
-
-user_schema = UserSchema()
-users_schema = UserSchema(many=True)
-
-db.create_all()
-
-@app.route('/user', methods=['POST'])
-def add_user():
-    name = request.json['name']
-    email = request.json['email']
-    password = request.json['password']
-
-    new_user = User(name, email, password)
-
-    db.session.add(new_user)
-    db.session.commit()
-
-    return user_schema.jsonify(new_user)
-
-@app.route('/user', methods=['GET'])
-def get_users():
-    all_users = User.query.all()
-    result = users_schema.dump(all_users)
-    return jsonify(result)
-
-@app.route('/user/<id>', methods=['GET'])
-def get_user(id):
-    user = User.query.get(id)
-    return user_schema.jsonify(user)
-
-@app.route('/user/<id>', methods=['PUT'])
-def update_user(id):
-    user = User.query.get(id)
-
-    name = request.json['name']
-    email = request.json['email']
-    password = request.json['password']
-
-    user.name = name
-    user.email = email
-    user.password = password
-
-    db.session.commit()
-
-    return user_schema.jsonify(user)
-
-@app.route('/user/<id>', methods=['DELETE'])
-def delete_user(id):
-    user = User.query.get(id)
-    db.session.delete(user)
-    db.session.commit()
-
-    return user_schema.jsonify(user)
-
-if __name__ == '__main__':
-    app.run(debug=True, env='development')
+    return app
